Engineering Task #3303 hbm file changes to support genericDao
git-svn-id: http://mcoltrac/svn/mcol-app/sdt/trunk@16655 e0420d7e-bc29-8447-b031-85aca2f896c5

+++ domain/src/main/java/uk/gov/moj/sdt/domain/AbstractDomainObject.java
@@ -48,7 +48,7 @@ public abstract class AbstractDomainObject implements IDomainObject, IVisitable
     /**
      * Primary key.
      */
-    private int id;
+    private long id;
 
     /**
      * Hibernate version number.
@@ -76,13 +76,13 @@ public abstract class AbstractDomainObject implements IDomainObject, IVisitable
     }
 
     @Override
-    public int getId ()
+    public long getId ()
     {
         return id;
     }
 
     @Override
-    public void setId (final int id)
+    public void setId (final long id)
     {
         this.id = id;
     }
